+++ pkgs/ffigen/lib/ffigen.dart
@@ -23,6 +23,7 @@ export 'src/config_provider.dart'
         CommentType,
         CompoundDependencies,
         Declaration,
+        Declarations,
         DynamicLibraryBindings,
         EnumStyle,
         Enums,