+++ pkgs/ffigen/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 20.1.0
 
+- Add `xcodeUri`, `iosSdkUri`, and `macSdkUri`, to mirror `xcodePath`,
+  `iosSdkPath`, and `macSdkPath`.
+- Export some missing elements from the config API.
 - Provide more convenience utils for building FFIgen configs:
   `Declarations.excludeAll`, `Declarations.includeAll`,
   `Declarations.includeSet`, `Declarations.includeAllMembers`,