+++ pkgs/ffigen/CHANGELOG.md
@@ -3,6 +3,12 @@
 - Add `xcodeUri`, `iosSdkUri`, and `macSdkUri`, to mirror `xcodePath`,
   `iosSdkPath`, and `macSdkPath`.
 - Export some missing elements from the config API.
+- Provide more convenience utils for building FFIgen configs:
+  `Declarations.excludeAll`, `Declarations.includeAll`,
+  `Declarations.includeSet`, `Declarations.includeAllMembers`,
+  `Declarations.includeMemberSet`, `Declarations.useOriginalName`,
+  `Declarations.renameWithMap`, `Declarations.useMemberOriginalName`, and
+  `Declarations.renameMemberWithMap`.
 
 ## 20.0.0